fix: repo return optional value

+++ src/main/java/ru/tututu/trains/repo/ReservationRepo.java
@@ -3,6 +3,7 @@ package ru.tututu.trains.repo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import ru.tututu.trains.entity.Reservation;
+import ru.tututu.trains.exceptions.NotFoundException;
 import ru.tututu.trains.mapper.ReservationMapper;
 import ru.tututu.trains.utils.DataSourceProxy;
 import ru.tututu.trains.utils.params.IntegerParam;
@@ -10,7 +11,9 @@ import ru.tututu.trains.utils.params.QueryParam;
 import ru.tututu.trains.utils.params.StringParam;
 
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Component
 public class ReservationRepo {
@@ -38,6 +41,9 @@ public class ReservationRepo {
     }
 
     public void deleteReservation(int id) throws SQLException {
+        if(findReservationById(id).isEmpty())
+            throw new NotFoundException("reservation with id " + id + " not found");
+
         String sql = """
                 DELETE FROM reservation
                 WHERE reservation.id=?
@@ -45,5 +51,13 @@ public class ReservationRepo {
         QueryParam[] queryParams = new QueryParam[]{new IntegerParam(id)};
 
         dataSourceProxy.executeDelete(sql,queryParams);
+    }
+
+    public Optional<Reservation> findReservationById(int id) throws SQLException {
+        String sql = "SELECT * FROM reservation WHERE id=?";
+        QueryParam[] queryParams = new QueryParam[]{new IntegerParam(id)};
+        List<Reservation> reservationList = dataSourceProxy.executeSelect(sql, new ReservationMapper(), queryParams);
+
+        return reservationList.isEmpty() ? Optional.empty() : Optional.of(reservationList.get(0));
     }
 }
